Remove glob include_dotfiles param
Summary:
Buck v1 does not have `include_dotfiles` param.

IIRC it was removed from Buck v1 because it is too hard to make it work correctly in all four glob implementation in Buck v1. It is not an issue in Buck v2, but we don't really need this param: once in a lifetime when dotfiles are needed, these can be included with a pattern starting with dot:

```
glob(["*", ".*"])
```

`include_dotfiles` params is **rarely needed and redundant**, so I'm in favor of removing it.

Also [Bazel's glob](https://docs.bazel.build/versions/main/be/functions.html#glob) does not have `include_dotfiles` either.

Also there are no integration tests for `include_dotfiles` param.

But we can keep it.

Reviewed By: ndmitchell

Differential Revision: D36961420

fbshipit-source-id: 8170d22ab9fd2f26d1c7e3c97936da3dc8ee7c8a