Add env var to disable injected-key time-travel bugs
Summary: Details in comment - now that we have an answer key, the fuzzer is finding lots of places where we _should_ panic, but don't, since DICE is reading an injected-key from the future. This adds a temporary mode that doesn't search for "DICE didn't panic, but should have" errors, but still catches "DICE computed the wrong value" errors.

Reviewed By: krallin

Differential Revision: D36975654

fbshipit-source-id: b02e0d950632f2843430750a57867ed2e2197493