allow executable genrules with multiple outs
Summary:
D36679626 broke minimal xcode toolchain builds for buck 2 as it does not support genrules with multiple outputs being executable. This does not match the behaviour in Buck 1 where the default output would be the $(exe) and the remaining outputs would be materialized.

This diff updates the genrule in Buck 2 to match this behaviour.

Reviewed By: ndmitchell

Differential Revision: D36867649

fbshipit-source-id: 6b45013d885950b195d21d2fbb06d63f2fda3ea4