buck2: dep files: don't specifically expect a BuildArtifact
Summary:
We currently store a BuildArtifact as the output where a dep file is, rather
than any artifact. This is annoying for projected artifacts, but it's also
somewhat unnecessary.

Indeed, the only use case we have for this is verifying that the action
produced the depfile as an output, but this is:

- Unnecessary nowadays because we defer reading dep files until the action runs
  again so this would be reported as an error when the action returns.
- Not particularly reliable anyway: the output could be a directory and that's
  not more readable than a non-existent file!

So, let's clean this up.

Reviewed By: bobyangyf

Differential Revision: D36865871

fbshipit-source-id: 4e0e8f7ddd404cd5738144df27acd424fc38c2d8