buck2: Projected Artifacts: introduce them
Summary:
Like it says in the title, this adds support for "Projected Artifacts" in
Buck2, which means an artifact representing a path within another artifact.

Most of the complexity here comes from the fact that we need to pick a
representation for `Artifact` and it turns out there is not really one that
works for everything. We need to choose if artifact looks like this:

```
enum Artifact {
  Projected(BaseArtifact, ProjectedPath),
  Base(BaseArtifact)
}

enum BaseArtifact {
  Build,
  Source
}
```

Or like this:

```
enum Artifact {
  Build(BuildArtifact),
  Source(SourceArtifact)
}

struct SourceArtifact(Source, ProjectedPath);
struct BuildArtifact(Build, ProjectedPath);
```

As mentioned, there isn't one that works for everything.

In particular, the artifact computation code gets very awkward with the latter
and it's conceptually not quite right because it allows actions to produce
projected outputs.

On the flip side, the former representation (which I picked) is uncomfortable
in the Starlark side (this diff) because a number of places "close to Starlark"
want to deal with artifacts knowing they are build artifacts and the first
approach doesn't allow for representing "Build + Projected Path", so we need a
new struct for this (this is `BoundBuildArtifact` in this diff).

Other than this, this diff is mostly mechanical changes arising from:

- `get_path` can no longer return a borrowed path (because one might not exist)
  so now it returns a new dedicated stuct. Same with all the fingerprinting
  stuff.
- A bunch of Starlark places that wanted paths as reference need to be ready to
  receive a `Cow::Owned` instead.

Note that most places now transparently deal with projected artifacts, with
the exception of a few things:

- Outputs are *not* projected. This makes sense: actions are supposed to build
  the outputs not the projection.
- Some places materialize everything regardless of projections (same with unhashed symlinks). This
  could be optimized of course, but I didn't do it. Still, we materialize more
  than we need to but it's not incorrect, just inefficient (somewhat).

Reviewed By: ndmitchell

Differential Revision: D36865878

fbshipit-source-id: 48fed17002a29aa6b1b1ac6613cde71d2613217e