buck2: StarlarkArtifact: use the same `ArtifactPath` as everything else
Summary:
We have many ways of doing the same things in artifacts, I'm trying to
standardize it all (to the extent possible) in in `ArtifactPath`, which is
basically a reference type that we can produce from _anything_ that is
an artifact and use for meaningful comparisons, `Display`, or for helper
methods.

The only missing part here was source artifacts, which this adds.

Reviewed By: bobyangyf

Differential Revision: D36893406

fbshipit-source-id: 7531e769ad27c578da0994acc8e152d0972d5685