Deps-less nodes are not always re-usable, even though no deps have, technically, changed
Summary:
Querying an injected key that hasn't been set yet calls its `compute()` method and panics. However, when we query a key at a version for which it is Unknown, we check if its deps haven't changed for that version range - if none of the deps have changed, we know we can re-use the node.

The problem with this is that keys with no deps can never have their version change, and so any deps-less key will be re-used for all other versions - which is incorrect when we're querying at different versions.

Reviewed By: bobyangyf

Differential Revision: D36991499

fbshipit-source-id: dc6f6e3fa6498911c6a3d78940b35dbe7e9103fd