android_apk cpu_split_transition should account for cpu_filters
Summary:
Previously, if a `ovr_config//cpu/constraints:cpu` constraint was not set, then we would error out when analyzing an Android APK rule.

Now that we are able to access the `cpu_filters` attribute on `android_binary`, we are able to do the following:
- If a CPU constraint and the `cpu_filters` attribute is not empty, check that they match and do the split transition.
- Otherwise, do the split transition based on the one that is set.
- If neither are set, complain.

We could simplify things further by _only_ using the `cpu_filters` attribute. However, I decided against that because it's quite nice just to be able to use the configuration directly in the case where we only have one CPU specified. If people think that isn't worth it, we can remove it.

Reviewed By: stepancheg

Differential Revision: D36758636

fbshipit-source-id: 66dc961d28aa44d7981fa002115cf77f8c5f1979