Fix incorrect link style when building shared libs
Summary:
# Context

In Buck (1+2), all linkable cxx rules have two fields which control how linking works:

- [link_style](https://buck.build/rule/cxx_library.html#link_style): Determines whether to build and link *this rule's dependencies* statically or dynamically.
- [preferred_linkage](https://buck.build/rule/cxx_library.html#preferred_linkage): Controls how a library should be linked.
  - `any`: The library will be linked based on its dependents `link_style`
  - `shared`: The library will be always be linked as a shared library.
  - `static`: The library will be linked as a static library.

When a particular target needs to be linked (e.g., `(apple|cxx)_(library|binary)`), the `link_style` of the target is taken and applied down the dep-tree (i.e., `link_style` travels down and it's combined with `preferred_linkage` to determine linking).

In Buck2, we can only propagate information up, which means that dependencies *must* provide all possible outputs and then the dependents can decide what to use. When it comes to linking, we generate outputs for *all* link styles and pick the appropriate one when computing the output for shared outputs of libraries and executables for binaries.

# Problem

There was a bug in Buck2 which ended up producing incorrect linking. Specifically, we were *always* linking using `shared` link style and ignored the actual value provided by the target.

**Example**

Imagine we have the following dep graph: `A -> B` and the following attributes:

- `A`
  - `link_style=static`
- `B`
  - `preferred_linkage=any` (the default)

If we then build `buck2 build //:A[shared]`, we *must* link `B` statically because `link_style=static` (and that's exactly how Buck1 behaves).

But before this diff, we will ignore the `link_style` property

NOTE: Linking behaved correctly when using link-group linking but *not* otherwise - see how we lifted usage of the link style to apply to both types of linking.

# Solution

Use the link style for the target as per Buck1 behaviour and as intended by docs.

Reviewed By: andrewjcg

Differential Revision: D37107259

fbshipit-source-id: b625015c685b52ff6afb07251f124c804fadb5bc