Have cquery functions use more powerful targetexpr
Summary:
Allows more flexibility in api with bxl functions, so that we can pass results of queries, cli args, etc around easily into these functions.

Previously, this had to be a list of target node or a str, which wasn't always obtainable unless we called these functions to obtain it.

Reviewed By: stepancheg

Differential Revision: D37060629

fbshipit-source-id: b566d4e9d691f85fb19a0f2fc83f61b84246737f