Make HARD_ERROR bubble up properly with soft_error!
Summary: We want to write tests of soft_error! as though they give back errors. We first did that with panic, then with an exit, but in both cases we end up fighting behaviours that won't be the end state after we make them real errors. Fix that by actually propagating a real error with `?` if you have HARD_ERROR set.

Reviewed By: krallin

Differential Revision: D37107689

fbshipit-source-id: da70dc8817fcd3e0fad7b0486335069804180796