Add a `$package` attribute to `buck targets` output
Summary:
This gives us enough information from `buck2` to be able to construct the full target graph.

## Why?

I want to figure out whether or not we can build up the `buck2` target graph using `$deps` and dropping non-existent edges. That way, we can make the TD process more resilient to missing nodes.

Rough sketch

```
$ buck2 targets --output-attribute '^(\$deps|\$package|name)$' //...
```

to parse out the target name (`$package:name`) and all of its `$deps`. Importantly, `$deps` can be computed and returned without resolving the target graph completely, which means we are able to handle missing targets and missing edges.  Eventually, it might be nice to be able to do the entire graph pruning and construction in rust (I think?) though we'd need a different type of command (or a `--keep-going` flag to continue parsing even in the face of an error).

Reviewed By: cjhopman

Differential Revision: D37109489

fbshipit-source-id: 8f9012d21b61d545cc73af66c5cc7ed6465f0ae8