buck2: deferred materializer: accept prefixes on all codepaths
Summary:
We had an inconsistency between the method that materializes and the one that
tells you what paths need materializing. The former accepts only exact
matches, but the latter accepted prefixes. This diff updates both to
allow prefixes (this is useful with projected artifacts).

I also updated the "finish" method to stop requiring `entry`. It used to need
this to delete things out of the materializer state, but we've stopped doing
this, so we no longer need it.

Reviewed By: scottcao

Differential Revision: D37134771

fbshipit-source-id: 86aa20c9129697a0561b9e2c163e1cad1dea991c