Add linkable for Apple Frameworks
Summary:
Currently, the FBiOS is blocked from building on `remote/master` because it's argsfile is filled with duplicated `-f` and `-framework` linker args. This stack works to dedupe them.

The first step is adding a new `LinkableType("frameworks")` to the graph. One introduced we will have a extra step to dedupe and expand the `LinkableNodes` -> `[""]` outside of the usual linker_flag creation process.

Reviewed By: milend

Differential Revision: D36989659

fbshipit-source-id: 6b961f3def5cfdc009eba620c1be180d78369b55