Stop materializing the same artifact if it's already materialized
Summary:
This diff checks the existing materializer state when a new artifact is declared. If that artifact is the same as the existing artifact already materialized, then we skip the deletion and declare steps. We currently check whether the artifact is same by checking equal on the ArtifactValue.

This feature is currently gated behind a buckconfig `buck2.enable_local_caching_of_re_artifacts`, in case there are unforeseen cases where we cache too much that would require a rollback.

Reviewed By: krallin

Differential Revision: D36790791

fbshipit-source-id: f408becf8d4c9c000cf3f60840c1f14b0e5fc894