Make local executor clean up its own output directory
Summary: Right now, local executor does not declare outputs on the materializer. This is problem because it means that it won't be able to take advantage of the materializer change to do delete-on-declare. Here, we add a temporary workaround by making local executor explicitly clean up its own output directory. We also thread in the outputs cleanup bool from incremental API to make sure that we only cleanup outputs if that's requested.

Reviewed By: krallin

Differential Revision: D36789035

fbshipit-source-id: 7af38afb2d67efc4d7d4655c4595bf746e97e78d