Make declaring an artifact responsible for deletion
Summary:
In order to have a local action cache, we must make sure that the old artifact, if it exists, is not deleted when an artifact with same path is declared to the materializer. This means that materializer's declare must be in charge of whether to delete this artifact. This diff adds artifact deletion to the declare step of the deferred materializer.

Now the declare step will spawn a future that deletes any conflicting path. This future will be stored in Declared stage. When the artifact is requested to be materialized, we pass this deletion future to the spawned materialization future. The materialization future first waits on the deletion future to finish before it tries to materialize any artifact.

Reviewed By: krallin

Differential Revision: D37042322

fbshipit-source-id: 65127b677bf261347ec47b9b9ccdf19d6520e1cb