Mark python_binary targets as incompatible
Summary:
python binaries (at least in-place ones) encode an absolute path to an
interpreter that may not be cross-platform and so we need them to be marked
fat-platform incompatible.

Ideally, the toolchain could be a target dep (instead of exec dep) and then the
interpreter could be a target dep and then the incompatibility could be moved to
the interpreter (and, importantly, if the interpreter were actually x-plat, it
could be compatible).

Reviewed By: IanChilds

Differential Revision: D37125705

fbshipit-source-id: 855270c31ea52491afbfd7bbdba0cacfa6c03fcc