Prevent optimizer running very long on certain inputs
Summary:
Consider this example:

```
def a(): return "a"
def b(): return a() + a() + a() + a() + a()
def c(): return b() + b() + b() + b() + b()
...
def z(): return y() + y() + y() + y() + y()
```

Without this diff, evaluation of this code runs for ages and takes a lot of memory, even if neither of functions is called. Because it is optimized into:

```
def a(): return "a"
def b(): return "aaaaa"
def c(): return "a..." # 25
...
def z(): return "a..." # 298023223876953125
```

This is not practical issue in our setup, but probably better if we avoid issues like that.

Reviewed By: ndmitchell

Differential Revision: D36712593

fbshipit-source-id: f1b5ff5c710eec8d312c91bba2f64f363cb276c4