Give prebuilt_jar its own toolchain
Summary:
I'm about to add a tool to the Java toolchain that is written in Kotlin.

As it is, doing so will create (many) cycles from the toolchain to itself. The ideal solution (D37130054) can't be done right now because it would only remove the cycles in the *configured* target graph. Leaving cycles in the unconfigured graph would likely break downstream tools that aren't expecting that. Instead, I'm going to have to go visit all of the transitive deps of the tool and switch them over to using a bootstrap toolchain.

Many of those dependencies are `prebuilt_jar`s. `prebuilt_jar` is kind of an outlier in the world of Java rules in that it uses very little of the toolchain (just the ABI generator). Accordingly, this diff gives it its own tiny toolchain that thus won't be involved in cycles with new tools in the main Java keychain.

Reviewed By: IanChilds

Differential Revision: D37178465

fbshipit-source-id: d3a1aa3d54a2553286b7a0faf4c951dd6f6c7f62