Only materialize the header map rather than full build
Summary:
This should ensures the materialization of "-I" hmaps in the compilation database. I speculate that it would be enough for clangd.

I'm going to add this in for now, turn on bxl for testing next week and see if we get any problem with it. I left some TODO there in case there is a problem.

Reviewed By: etriebe, bobyangyf

Differential Revision: D37227440

fbshipit-source-id: bb45ba71358184ba9a13afe73254392a513e48f6