buck2: prelude: needed_coverage: match unqualified source names
Summary:
For explicit modules, we currently look for the module name that was passed in,
but the array we look into contains fully qualified names, so that doesn't
work out.

We still need the fully qualified names when creating the command line (that's
what Buck1 does at least), so update PythonCoverageInfo accordingly.

Reviewed By: andrewjcg

Differential Revision: D37234723

fbshipit-source-id: 3ef5d5f8aff540ff10731f182f4952fb37267345