buck2: test_build_isolated move this to a unit test
Summary:
This regex matching is pretty brittle and a pain to iterate on. I run into it
every time I touch the superconsole, so I'd like to move this to a place where
it's a) hopefully less brittle and b) doesn't require actually building buck2
and running it to run the assertions.

Note that this doesn't check the final console (because we can't easily access
the final output), but we have other unit tests for `DrawMode::Final` on
individual components so that seems less necessary.

Reviewed By: ndmitchell

Differential Revision: D37276448

fbshipit-source-id: 9951c3cee28f4d35c32871c432563220445d763e