Fix Clang modules when compiling using toolchain
Summary:
# Problem

When using `apple_toolchain()`, we use a relative path to `resource-dir`. The compiler will then generate `#include` paths relative to fbsource and those won't work without additional directives.

One way to solve would be to absolutise paths in a wrapper but that will end up producing abs paths in cacheable outputs which is undesirable.

# Solution

Use `-I.` to make the generated includes work.

Reviewed By: rmaz

Differential Revision: D37242799

fbshipit-source-id: 7a9f15ec26b54991e79811ae7a37e02daa5b0354