buck2: test: send structured target data to Tpx
Summary:
The cell inference in Tpx right now is somewhat broken on v2 (it infers a
lot of things are fbcode when they shouldn't). To make this more robust, Tpx
would like us to send more structured data.

This does it. We can eventually retire the `name` here once Tpx migrates
off of using it.

Reviewed By: ndmitchell

Differential Revision: D37307984

fbshipit-source-id: 4f595a6fbdcdd5c1c194b6dc4a1595188c9a4497