buck2: flush dep files on fresh instance
Summary:
We sometimes see errors due to dep files no longer existing that appear to be
caused by the user clearing out some files. We don't see this as often for
builds, and one difference is dep files aren't flushed when we get a fresh
instance.

I'd like to start clearing dep files on fresh instances so that we can get
better visibility into whether there is a bug in dep files specifically or not.

Note: there is some potential performance degradation here obviously, but since
our perf on a fresh instances is already trash and dep files weren't designed
to solve that particular problem, I think we should be OK with this.

Reviewed By: ndmitchell

Differential Revision: D37311146

fbshipit-source-id: a4ee262e0da3421ba20b73a0ea81d188b77711db