Resolve string literals in buck lsp context
Summary:
Implement resolve_string_literal() for buck. This looks for the file to be a target first, and if it's not, then it tries to resolve it as a relative path. If that doesn't work, no URL is returned.

For targets, we also look for a calling function in the buildfile of the package that has a "name" kwarg that matches the target. If it's not found, we just return (0, 0) for the location.

Reviewed By: ndmitchell, stepancheg

Differential Revision: D37298748

fbshipit-source-id: 6f50db5acb39bc325f84d97d64cef16e8d7ac1c0