Add "ast" output that dumps the AST of the Java code that was compiled
Summary:
This diff adds a new output to all `java_library` (and descendant) rules that produces a serialized AST for the code in that rule. In the process it changes a few targets to build using the bootstrap toolchain to avoid target graph cycles.

The AST-dumping plugin is still in development. At the moment it dumps literally just the AST, but the plan is for it to expand to dump type references as well.

Reviewed By: IanChilds

Differential Revision: D37130052

fbshipit-source-id: 298c25ab047ef0910f01490affd0697c6d6fa28a