Inplace par bootstrapping can run remotely
Summary:
# Context

We're observing significant regression in cold builds for Messenger using Buck2 (in the order of taking around 50% more time). This was traced due to running par bootstrapping rules locally (see T124153284).

# Solution

In-place par bootstrapping can be cached as it doesn't need to be build stamped (i.e., output will be deterministic).

Reviewed By: IanChilds

Differential Revision: D37344135

fbshipit-source-id: 062ee40315eefb5692b18ff21bfb5e84542e9948