Make sure we do full text match for link group label
Summary:
# Context
While getting fbios link group build to work, I noticed there is a behavior inconsistency between buck1 and buck2 link group label match, where in buck1 link group label matches the full text with an expression, in buck2 the expression can match anywhere in the text.
For example,
public.profile(\\..*)? and public.profileplus is a no match in buck1, but matches in buck2.
This inconsistency causes fbios-link-group build to break with buck2.

# Proposed Fix
To fix it, add anchors to the passed-in link group label regex to ensure that full text matches an expression.

Reviewed By: milend

Differential Revision: D37346126

fbshipit-source-id: 17e55be0b0cab8bbd5e2795c43b8aa72ec5658ba