Use FrozenStringValue in module names
Summary:
Instead of `String`.

Allocate strings only once.

This does not have performance effect, but it is safer this way.

Following diff converts compiler to use `FrozenStringValue` for identifiers instead of `String`, for similar reason.

And in the top of the stack D36401234 bytecode is printed with variable names instead of slot numbers.

Reviewed By: ndmitchell

Differential Revision: D36400429

fbshipit-source-id: 8f008408b9263bffd37938eb5ceeafb0aa945b47