buck2: move malloc_stats_print to buck2 internal crates
Summary:
I put this in common/rust in D30930750 8 months ago, but we're still the
only users of this, so it's probably not that useful. Also, unlike the rest of
this crate this doesn't use folly::memory, so it's probably not that an idea
anyway.

Actually porting this to Folly seems like it'll take some time since I'm not
getting much traction on even discussing how to approach this
(https://fb.workplace.com/groups/560979627394613/permalink/2409816815844209/),
so for now I'd like to take this back and put it in buck2 where it won't need
to be as general.

Reviewed By: ndmitchell

Differential Revision: D37348286

fbshipit-source-id: e052f19dba4ac2bd074fb6f0e2118c184dee24d9