buck2: do not create dirs in CleanOutputPaths
Summary:
Like it says in the title. This creates a bunch of I/O for paths we might not
actually need (if we run those actions on RE and never materialize their
outputs).

If it turns out to be needed then we'll create the paths as needed. We already
do the parent dir creation in:

- LocalExecutor
- Writing files
- Creating copies
- `materialize_dirs_and_syms`

Those are, I think, all the places that depend on this.

Reviewed By: ndmitchell

Differential Revision: D37106560

fbshipit-source-id: 4ed02e79307350ee23f749ed95b5e0ee08388379