START REVIEW HERE [buck2] Move ConfiguredAttrTraversal to buck2_node
Summary:
Enough for today.

I think it is possible to have a crate `buck2_node`, which contains `TargetNode`, `ConfiguredTargetNode`, attr types, and basic attr logic.

Up to this diff, `AttrType` and `ConfiguredAttr` is moved into `buck2_node` crate. More work is needed to move the rest. But it seems doable.

`buck2_node` crate intentionally does not depend on `starlark` crate (explicitly or transitively).

Upside is `buck2_node` is typechecked or compiled concurrently with `starlark` crate.

Downside is that coercion and resolution of `Value` to `AttrLiteral` lives in `buck2_build_api` crate.

However:
* we can always add dependency on `starlark` (and adding a dependencies is much easier to removing dependencies)
* or we can put coercion into another crate like `buck2_node_coerce` which would depend on `buck2_node` and `starlark`

With this diff, the size of `buck2_node` crate is 1800 LOC (including license headers). Size of `buck2_build_api` is 53000 LOC. It will grow further when `TargetNode` is moved to `buck2_node`, so size of that crate won't be groundbreaking for `buck2_build_api`, but it should be significant.

Reviewed By: bobyangyf

Differential Revision: D37370430

fbshipit-source-id: 2e079e3e33a0060d5ee60fd493f45966afdb5f53