Change get_template_info_provider() to return artifacts
Summary:
We need to traverse tsets for these lookups and to do that we need to
handle `ArtifactGroup::TransitiveSet`. Now, two things are common: (1) we get
a call with a large TargetSet in `get_from_template_placeholder_info` and (2)
the tsets that we're going to traverse across multiple of those targets are in
the same tset-graph.

This means that we need a single traversal over the full set of tset nodes rather
than traversing them for each target individually and so we propagate the artifacts
up to where we can see the full set of targets.

Reviewed By: stepancheg

Differential Revision: D37368878

fbshipit-source-id: 4670facca0a7b3a3f54943f56640d933e25cab5e