Change ConfiguredTargetNode to hold direct references to deps
Summary:
This will allow us to implement the analysis-time query environment on
the ConfiguredTargetNodes directly (right now it produces a completely separate
graph to operate on).

This also reduces memory use for the target graph since ConfiguredTargetLabel
is significantly larger than Arc. For m4a, this reduces configured graph memory
use from ~2GB to ~1.7GB (note though that the unconfigured graph holds most
of the data and that's ~7.6GB there).

Reviewed By: scottcao

Differential Revision: D37335713

fbshipit-source-id: 208960311a0838082cb1ae31a95886c2acb08a82