Fix shutdown handling in stdinstream
Summary: Previously we were only checking the shutdown oneshot set in Drop if stdin isn't "ready". Turns out sometimes after you write to stdin the first time, it will always report back "ready". While that's its own issue, fix this so that we always check if we've been told to shutdown so that we don't hang on ctrl-c.

Reviewed By: bobyangyf

Differential Revision: D37399904

fbshipit-source-id: 5553a77ad8bcf8f74e08cde94c095b168836845c