compute() functionality extended to also hash UnconfiguredTargetNodes
Summary: The compute() function of TargetHashes previously only worked for computing hashes of ConfiguredTargetNodes, but now has a generic type parameter that specifies whether to run target hashing on (Unconfigured)TargetNodes or ConfiguredTargetNodes. The flag --show-target-unconfigured-hash introduced in a previous diff allows the user to run unconfigured target hashing from the command line.

Reviewed By: stepancheg

Differential Revision: D37301601

fbshipit-source-id: baf4b8b60275035817f43ab7559c97898a1cc977