Return a .idea directory containing the full project
Summary:
Right now we return the path to the `modules.xml` file. Instead, we want to return the path to the `.idea` directory, and the output of this command should be the directory. Therefore, make the `.idea` directory contain a symlink tree to all the outputs.

This also has the nice property that we only have to ensure the `.idea` directory exists and everything will get written as we expect.

Ultimately, we might want to find a different way to write everything to the `.idea` dir (e.g. one action that does all the writing), but for now this will do.

Reviewed By: cute-jumper

Differential Revision: D37414150

fbshipit-source-id: a9434397e9e702017729d1e402f81b92d7339b3c