buck2: python: always link C++ libs against libomnibus
Summary:
This mechanism looks like it didn't quite work when it got implemented,
because we have a bunch of body nodes that always cause the omnibus.so to be
added, like those:

```
fbcode//third-party-buck/platform009/build/glibc:libc (ovr_config//platform/linux:x86_64-fbcode-platform009-clang-136d45797747f845)
```

In platform010 those appear to be gone and that causes libraries to not link
against libomnibus.so and breaks tests that exercise this exact behavior
(`tests/targets/rules/python/omnibus/mult_cpp_ext_roots`).

Since the behavior of omitting libomnibus (presumably for perf reasons) doesn't
look like it has test coverage, but the behavior of ensuring symbols provided
by other deps are visible *is*, I'm opting to keep the latter and remove the
former.

Reviewed By: ndmitchell

Differential Revision: D37453074

fbshipit-source-id: 01af80ce1f7698220d1993a455f4b52fda67a192