java_test runs from project_root unless test specifies that it should run from the cell root
Summary:
We want to be able to run `java_test`s from the project root so that when we use `--unstable-allow-tests-on-re` the tests do actually run on RE.

Right now they run from the cell root. One option that would work for us - since we want to run tests on RE in fbsource (and would not be potentially breaking for users, unlike this diff) would be to set `run_from_project_root = True` and `use_project_relative_paths = True` when the cell root matches the project root. However, I didn't see a way to get the project root in order to compare it with the cell root.

Instead, just requires users to opt into running from the cell root.

Reviewed By: krallin

Differential Revision: D37380832

fbshipit-source-id: 7dc233ab96951baf96a3cc708a91de66e1a37c15