Add basic support for prebuilt_jar
Summary:
`prebuilt_jar` is represented in the Intellij Project as a "Library", per https://www.internalfb.com/intern/wiki/Developing-intellij-plugins/IntelliJ_Project_Generation/Library/.

In this diff, we just assume that all `prebuilt_jar`s are "Project Libraries", and so we write them to their own file `.idea/libraries/`.

To do that we iterate through all the target nodes, and if the `rule_type` is `prebuilt_jar`, then we create an `IJLibrary` for them. We then use that `IJLibrary` to write out the project library file.

There are a bunch of TODOs in here for things that I wasn't able to implement properly given the APIs as they currently are (or at least my understanding of them). bobyangyf, what's the best way to keep track of all these asks? I guess I'll just create tasks for each of them?

Reviewed By: cute-jumper

Differential Revision: D37416479

fbshipit-source-id: bf3816f1d0062ce51849cd9b330a22343b9a4ddc