Add very basic java_library support
Summary:
Really, I just wanted to test out using `bxl` and also get a bit more familiar with project generation. This diff checks if a node is a `java_library`, and if it is then it creates a simple data structure containing the deps (this comes from the `compiling_deps` so includes all `exported_deps` automatically, which is kind of nice), the `file_name`, the `rule_type` and the `srcs`. It then writes a really simple `.iml` file for each module that it created.

This is missing "merging", which I'll probably try to add in a future diff, as that adds a bunch of complexity and we need to see how that shakes out.

Reviewed By: cute-jumper

Differential Revision: D37380560

fbshipit-source-id: fa504fa4a1f846406d4644817b3761dd2e3d4737