Tests can access private interface of direct deps
Summary:
# Problem

In Buck1, if a test target is listed in the `tests` field of a library, the test target has access to the *private* (i.e., non-exported) headers of said library.

This behavior is currently missing in Buck2.

# Buck1 Behavior

The behaviour comes from the following piece of code in Buck1 where we add the private headers of all direct deps of an `(apple|cxx)_test()`.

https://www.internalfb.com/code/fbsource/[e29d725377b3f92562bb9637766b04e8a1cdb216]/xplat/build_infra/buck_client/src/com/facebook/buck/cxx/CxxDescriptionEnhancer.java?lines=522-530

# Solution

We create a new provider `CPreprocessorForTests` which contains two pieces of information:

- The test targets (i.e., the only targets that are allowed access to the non-exported headers)
- The non-exported headers

When compiling tests, we then include the non-exported headers if we're allowed access.

To hook into all the relevant places, we change `cxx_private_preprocessor_info` to return a tuple of the private headers and any private headers of first order deps if the current target is a test.

- Q: Why not use a tset?
- A: `CPreprocessorForTests` is not propagated up the tree, only one level up and there's no merging/deduping.

Reviewed By: chatura-atapattu

Differential Revision: D37486045

fbshipit-source-id: 0375ee211ec3c935a276fbbde99ade4ad57052b7