Fix dep types of swift_toolchain() fields
Summary:
# Problem

As part the Swift Explicit Modules support, the `swift_toolchain` itself needs to execute the Swift tools part of the toolchain.

https://www.internalfb.com/code/fbsource/[eab46b45f4a68c62039092050972014852ab01e6]/fbcode/buck2/prelude/apple/swift_toolchain.bzl?lines=5

Consequently, this means that the `swift_toolchain` target itself will go through execution platform resolution. But since the Swift tools were *not* marked as `exec_dep`s, that means the resolution process will just pick the first default platform, which happens to the Fat Mac Platform. So, we ended up sending Mach-O executables to a Linux RE machine.

# Solution

Correctly set the type of dep to `exec_dep`, so that exec platform resolution works correctly.

Reviewed By: rmaz

Differential Revision: D37487716

fbshipit-source-id: 7125a8dfe03187ad72cde2045206bcedf9133368