generate CompDb for header only library with C++ compiler info
Summary:
Header only libraries (see an example below) have always been a pain for C++ LSP, because it does not generate the compilation database. To deal with that, we used to
1. Find the target that owns the header
2. Find a dependency that has a srcs attribute and depend on the target in 1. This process involves multiple rdeps query and is very expensive
3. Build the CompDb for 2, rewrite its entry to hopefully make it work for clangd

Example header only library:
```
cpp_library(
    name = "cpu_observer",
    headers = ["CpuObserver.h"],
    exported_deps = [
        "//admarket/adindexer/utrace:adindexer_utrace",
        "//admarket/adranker/post_ranking:post_ranking_stages",
        "//common/time:time",
    ],
)
```

This should enable the compDb generation for these libraries and reduce the query call # to 1. Also this makes our compDb generation process independent of rdeps query, which are not very reliable and can throw because of various reasons.

In this change, I picked C++ compilers for such kind of libraries, as there is no way to tell if the library is to be used with C or C++ sources. In the repos that Buck2 support right now, I have not found a header only library that's used in C source files. This is something to test when Buck2 is supported in more repos (e.g. firmware).

Reviewed By: williamtwilson

Differential Revision: D37477974

fbshipit-source-id: 6db32dc10a4a3ea699cc8b1be936f5251520ab79