Move async cleanup guard from run_streaming to cli::exec
Summary:
krallin noted on D37293872 that this is probably a reasonable place for async cleanup.

This diff adds a Guard type for AsyncCleanupContext whose drop implementation creates a runtime (since we can no longer share the runtime created in `run_streaming`) and joins on the scheduled tasks.

My plan for next week is to:
1. In `buck2/cli/src/lib.rs:exec`, store the exit code we get from `opt.exec(...)` in a global OnceCell.
2. Rebase D37293872 atop this.
3. Add a Drop implementation to the event-subscriber from D37293872, which schedules a cleanup handler that (when run) reads from the OnceCell storing an exit code, and fires a oneshot event to Scribe.

Reviewed By: krallin

Differential Revision: D37433834

fbshipit-source-id: eb1f40212b73507cd0ef6ae5684d95db8a2e2611