Make template_info based query funcs support tsets
Summary:
This is a bit tricky due to the inability to traverse cmdargs-like
things directly (we need to use visit_artifacts and that gives us
ArtifactGroup, cmdargs-like has no apis that can support the Value lifetimes).

So, we do two levels of traversal. For things that are opaque cmdargs-like (the
template_info value and the values contained in tset nodes) we extract out the
ArtifactGroups and then for any tsets in there we can lookup the tset and traverse
it directly (extracting out new ArtifactGroups from its values).

Reviewed By: krallin

Differential Revision: D37368875

fbshipit-source-id: 15f0fec49d6c7c5da65b322ad0857d9fffe1678b