buck2: update the consoles to be ready to receive CommandExecutionError
Summary: Like it says in the title. This is going to be the way we tell the CLI that a command failed (but that there is no _real_ error except for the fact that a command failed), in which case the CLI will look at the command reports (though we would look at the command reports anyway) to provide an explanation of what happened.

Reviewed By: ndmitchell

Differential Revision: D37488007

fbshipit-source-id: bf365777e5796757e6ff66480db3878879170575