buck2: data: report all command executions in `commands` field
Summary:
We want to be able to surface retries. To do this let's report all commands in
ActionExecutionEnd. The code changes are mostly straightforward serialization
here, but since we need to retain the old code (but don't want to e.g. fetch
stderr twice) there is a bit of refactoring here.

Push safety: this change only adds data, so it can go out without dependencies,
but it should land with the next diff to avoid logging too much.

Reviewed By: ndmitchell

Differential Revision: D37488005

fbshipit-source-id: ddeb3ca18239cacd335981a19826b34443180078