Make sure the manifest on an android_binary uses (one of) the same configuration(s) as the deps
Summary: Right now, these rules don't do a transition, so their entire dependency tree (which can be essentially the entire dep tree of the APK) is evaluated with the original configuration, whereas the deps of the APK are evaluated using (one or more) transitioned configuration(s). That's wasteful - instead we can just reuse the analysis that is already being done for the deps in one of their configurations (we just pick the first one).

Reviewed By: cjhopman

Differential Revision: D37457005

fbshipit-source-id: 8899e39c5c323a04902585995ee0c61384f45721