Add attr.toolchain_dep
Summary:
Prior to this diff every node has a single configuration (either exec or target), and the exec platform of a rule is determined by its immediate exec dependencies. That meant that the selection of a C++ toolchain had to be inlined into each and every cxx_library.

We now introduce toolchain_deps. These nodes don't select their execution platform, but instead have it forced on them by whatever includes them, hence it must be recorded in the configured target label. The execution platform resolution sees through them. Now (see 2 diffs later) the toolchain can be a named node, and referenced.

There are two immediate motivations for this work:

1. Save memory - 10Gb on targets, and will increase as we add more toolchains.
2. Open source compatible - we no longer need to list the toolchain in the Prelude.

Longer term this also gives us greater freedom to express kinds of toolchains we can't do today. But that's only a secondary benefit.

Reviewed By: cjhopman

Differential Revision: D37143042

fbshipit-source-id: 4f0ac86afc44a60c4a37bf24c20aaadd9a318145